#1363 Rename transaction.xa.manager.property => transaction.xa.manager.dialect
