+++ sharding-transaction/sharding-transaction-xa/src/test/java/io/shardingsphere/transaction/xa/manager/AllManagerTests.java
@@ -19,7 +19,7 @@ package io.shardingsphere.transaction.xa.manager;
 
 import io.shardingsphere.transaction.xa.manager.atomikos.AtomikosTransactionManagerTest;
 import io.shardingsphere.transaction.xa.manager.extractor.DataSourceParameterFactoryTest;
-import io.shardingsphere.transaction.xa.manager.property.XAPropertyFactoryTest;
+import io.shardingsphere.transaction.xa.manager.dialect.XAPropertyFactoryTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
