+++ sharding-transaction/sharding-transaction-xa/src/test/java/io/shardingsphere/transaction/xa/manager/dialect/XAPropertyFactoryTest.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.xa.manager.property;
+package io.shardingsphere.transaction.xa.manager.dialect;
 
 import io.shardingsphere.core.rule.DataSourceParameter;
 import org.junit.Before;
